requirements.txt
backcall==0.2.0
bokeh==2.1.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
idna==2.9
ipython-genutils==0.2.0
jedi==0.17.1
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
packaging==20.4
pandas==1.0.5
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.9
wcwidth==0.2.4
